+++ lib/razor/matcher.rb
@@ -16,6 +16,7 @@ require 'json'
 # The builtin operators are (see +Functions+)
 #   and, or - true if anding/oring arguments is true
 #   =, !=   - true if arg1 =/!= arg2
+#   like    - true if arg1 =~ arg2 (interpreting arg2 as Regex)
 #   in      - true if arg1 is one of arg2 .. argn
 #   fact    - retrieves the fact named arg1 from the node if it exists
 #             If not, an error is raised unless a second argument is given, in
@@ -62,6 +63,7 @@ class Razor::Matcher
         "tag"      => {:expects => [[String]],        :returns => Mixed   },
         "state"    => {:expects => [[String], [String]], :returns => Mixed   },
         "eq"       => {:expects => [Mixed],           :returns => Boolean },
+        "like"     => {:expects => [[String], [String]], :returns => Boolean },
         "neq"      => {:expects => [Mixed],           :returns => Boolean },
         "in"       => {:expects => [Mixed],           :returns => Boolean },
         "num"      => {:expects => [Mixed],           :returns => Number  },
@@ -123,6 +125,10 @@ class Razor::Matcher
 
     def eq(*args)
       args[0] == args[1]
+    end
+
+    def like(*args)
+      !(args[0] =~ Regexp.new(args[1])).nil?
     end
 
     def neq(*args)
@@ -304,6 +310,16 @@ class Razor::Matcher
                     "%{expected} are accepted") %
             {arg: arg.inspect, type: arg.class, name: name, position: pos, expected: expected_types}
         end
+      end
+    end
+
+    if name == 'like' and errors.empty?
+      begin
+        Regexp.new(rule[2])
+      rescue RegexpError => e
+        errors << _("invalid regular expression supplied to `like` for argument %{position}: %{message}") %
+            { position: caller_position || 1,
+              message: e.message }
       end
     end
   end