+++ spec/matcher_spec.rb
@@ -132,6 +132,28 @@ describe Razor::Matcher do
       match("=", "[]", "[]").should == true
     end
 
+    describe "like function" do
+      [['abc', 'abc'], ['abc', 'a.c'], ['abc', 'a.+'], ['abc', 'a.*'],
+       ['abc', 'ab'], ['abc', ''], ['abc', '\Aabc\z'], ['abc', '^abc$']].each do |str, reg|
+        it "matches with #{str.inspect} and #{reg.inspect}" do
+          match("like", str, reg).should == true
+        end
+      end
+      [['abc', 'def'], ['abc', 'z'], ['3', '#{1+2}'], ['abc', '\Abc\z'],
+       ['abc', '\zabc\A'], ['abc', '$abc^']].each do |str, reg|
+        it "fails with #{str.inspect} and #{reg.inspect}" do
+          match("like", str, reg).should == false
+        end
+      end
+      it "fails with invalid regular expression '*'" do
+        m = Matcher.new(["like", 'abc', '*'])
+        m.should_not be_valid
+        m.errors.should == [<<-EOF.strip]
+          invalid regular expression supplied to `like` for argument 1: target of repeat operator is not specified: /*/
+        EOF
+      end
+    end
+
     it "neq should behave" do
       match("!=", 1, 1).should == false
       match("!=", 1, 2).should == true
@@ -226,6 +248,14 @@ describe Razor::Matcher do
       Matcher.new(["=", true, false]).should be_valid
       Matcher.new(["eq", 1, ["=", 5, "ten"]]).should be_valid
       Matcher.new(["eq", 6.3, 3]).should be_valid
+    end
+
+    it "should allow strings for 'like' function" do
+      Matcher.new(["like", true, false]).should_not be_valid
+      Matcher.new(["like", 1, 2]).should_not be_valid
+      Matcher.new(["like", "abc", false]).should_not be_valid
+      Matcher.new(["like", 1, "abc"]).should_not be_valid
+      Matcher.new(["like", "abc", "def"]).should be_valid
     end
 
     it "should allow all types for '!=' function" do